Moved all utility functions to seperate subpackage and moved deepsalience download to generated folder.
