+++ timbre_trap/datasets/AudioMixtures/FMA.py
@@ -1,5 +1,4 @@
-from .. import AudioDataset
-from ..utils import *
+from .. import AudioDataset, stream_url_resource, unzip_and_remove, change_base_dir
 
 import numpy as np
 import os
