+++ timbre_trap/datasets/AudioStems/MedleyDB.py
@@ -1,6 +1,5 @@
 from ..Common import MedleyDB
 
-import warnings
 import os
