+++ timbre_trap/datasets/Common/TRIOS.py
@@ -1,5 +1,4 @@
-from .. import AMTDataset
-from ..utils import *
+from .. import AMTDataset, stream_url_resource, unzip_and_remove, change_base_dir
 
 import os
