+++ timbre_trap/datasets/SoloMultiPitch/MedleyDB_Stem_Synth.py
@@ -1,7 +1,6 @@
-from .. import BaseDataset
+from .. import BaseDataset, stream_url_resource, unzip_and_remove, change_base_dir
 from ..Common import MedleyDB
 from . import MedleyDB_Pitch
-from ..utils import *
 
 import shutil
 import os
