+++ timbre_trap/framework/__init__.py
@@ -1,4 +1,3 @@
-from .utils import *
 from .objectives import *
 
 from .cqtwrapper import CQT
